Issue 7958: Applied versions support for Text and Images only
git-svn-id: https://svn.imcode.com/imcode/imcms/trunk@13874 bd66a97b-2aff-0310-9095-89ca5cabf5a6

+++ src/main/java/com/imcode/imcms/mapping/orm/HtmlReference.java
@@ -10,7 +10,7 @@ import javax.persistence.Table;
 public class HtmlReference {
 	
 	@Id @Column(name="meta_id")
-	private Long metaId;	
+	private Integer metaId;	
 
 	@Column(name="frame_set")
 	private String html;
@@ -23,11 +23,11 @@ public class HtmlReference {
 		this.html = text;
 	}
 
-	public Long getMetaId() {
+	public Integer getMetaId() {
 		return metaId;
 	}
 
-	public void setMetaId(Long metaId) {
+	public void setMetaId(Integer metaId) {
 		this.metaId = metaId;
 	}	
 }
