+++ src/main/java/imcode/server/document/textdocument/TextDomainObject.java
@@ -62,7 +62,7 @@ public class TextDomainObject implements Serializable, Cloneable {
 	private Long id;
 	
 	@Column(name="meta_id")
-	private Long metaId;	
+	private Integer metaId;	
 	
     /**
      * Text filed index in the document.
@@ -214,11 +214,11 @@ public class TextDomainObject implements Serializable, Cloneable {
 		this.id = id;
 	}
 
-	public Long getMetaId() {
+	public Integer getMetaId() {
 		return metaId;
 	}
 
-	public void setMetaId(Long metaId) {
+	public void setMetaId(Integer metaId) {
 		this.metaId = metaId;
 	}
